move and reorganize tests
