+++ ijzer/src/tests/mod.rs
@@ -1,0 +1,8 @@
+
+#[cfg(test)]
+mod basic;
+#[cfg(test)]
+mod unary;
+
+#[cfg(test)]
+mod binary;
